we_are_social_inc_reusable_app/requirements.txt
arrow==0.7.0
beautifulsoup4==4.4.1
Django==1.9.6
django-debug-toolbar==1.4
django-disqus==0.5
django-emoticons==1.1.2
django-forms-bootstrap==3.0.1
django-paypal==0.2.7
django-tinymce==2.3.0
funcsigs==1.0.2
mock==2.0.0
pbr==1.9.1
Pillow==3.2.0
python-dateutil==2.5.3
requests==2.10.0
reusable-blog-app==1.0.0
six==1.10.0
sqlparse==0.1.19
stripe==1.33.0
git+https://github.com/peterbristow/reusable_blog_app.git
